+++ src/misc/validators.ts
@@ -1,5 +1,6 @@
 import {AbstractControl, ValidationErrors, ValidatorFn} from '@angular/forms';
-import {isBlank} from '@jscrpt/common';
+import {ModelPropertyMetadata, ValidatorFnFactory} from '@anglr/common/forms';
+import {isBlank, isNumber, isString} from '@jscrpt/common';
 
 import {DateApi} from '../services';
 import {DateTimeValueObject} from '../interfaces/dateTime/datetime.interface';
@@ -7,6 +8,35 @@ import {getSingleDateTimeValue, parseDateTime} from './utils';
 import {DateTimeValueFormat} from './enums';
 import {DateTimeInputOutputValue} from './types';
 import {DateTimeBase} from '../modules/dateTime/directives/dateTimeBase';
+import {DATE_API} from './tokens';
+
+//TODO: add decorators for min and max date time
+
+/**
+ * Validations arguments for date time validators for model based forms
+ */
+export interface DateTimeValidationArgs<TDate = unknown>
+{
+    /**
+     * Format of validated value
+     */
+    valueFormat?: DateTimeValueFormat|null;
+
+    /**
+     * Format of string value
+     */
+    stringFormat?: string|null;
+
+    /**
+     * Max allowed value
+     */
+    maxValue?: string|number|TDate|Date|DateTimeBase<TDate>;
+
+    /**
+     * Min allowed value
+     */
+    minValue?: string|number|TDate|Date|DateTimeBase<TDate>;
+}
 
 /**
  * Date time validator factory function, creates validator for checking validity of datetime
@@ -51,6 +81,72 @@ export function datetimeValidator<TDate = unknown>(dateApi: DateApi<TDate>,
 }
 
 /**
+ * Factory function that creates validator function factory
+ * @param args - Static arguments for date time validator
+ */
+export function dateTimeModelValidatorFactory<TDate = unknown>(args: DateTimeValidationArgs<TDate>): ValidatorFnFactory<DateTimeValidationArgs<TDate>>
+{
+    return new ValidatorFnFactory<DateTimeValidationArgs<TDate>>(args =>
+    {
+        const dateApi = args.injector?.get(DATE_API, null);
+
+        if(!dateApi)
+        {
+            throw new Error('DateTime: missing DateApi! Please provide one.');
+        }
+
+        return datetimeValidator(dateApi, args.valueFormat, args.stringFormat);
+    }, args);
+}
+
+/**
+ * Sets date time validator, with default formats
+ */
+export function DateTime(): PropertyDecorator
+/**
+ * Sets date time validator, with default string format
+ * @param valueFormat - Format of validated value
+ */
+export function DateTime(valueFormat: DateTimeValueFormat): PropertyDecorator
+/**
+ * Sets date time validator, with default value format
+ * @param stringFormat - Format of string value
+ */
+export function DateTime(stringFormat: string|null): PropertyDecorator
+/**
+ * Sets date time validator
+ * @param valueFormat - Format of validated value
+ * @param stringFormat - Format of string value
+ */
+export function DateTime(valueFormat: DateTimeValueFormat, stringFormat: string): PropertyDecorator
+/**
+ * Sets date time validator to property on which is used
+ */
+export function DateTime(valueFormatOrStringFormat?: DateTimeValueFormat|string|null, stringFormat?: string|undefined|null): PropertyDecorator
+{
+    let valueFormat: DateTimeValueFormat|undefined|null;
+
+    if(isNumber(valueFormatOrStringFormat))
+    {
+        valueFormat = valueFormatOrStringFormat;
+    }
+
+    if(isString(valueFormatOrStringFormat))
+    {
+        stringFormat = valueFormatOrStringFormat;
+    }
+
+    return ModelPropertyMetadata(
+    {
+        validators: [dateTimeModelValidatorFactory(
+        {
+            stringFormat,
+            valueFormat,
+        })]
+    });
+}
+
+/**
  * Date time validator factory function, creates validator for checking validity of datetime max value
  * @param dateApi - Date api used for parsing date time
  * @param maxValue - Maximal date time value that should be used for validation against
@@ -58,7 +154,7 @@ export function datetimeValidator<TDate = unknown>(dateApi: DateApi<TDate>,
  * @param stringFormat - Optional string format of value
  */
 export function datetimeMaxValidator<TDate = unknown>(dateApi: DateApi<TDate>,
-                                                      maxValue: string|number|TDate|Date|DateTimeBase<TDate>,
+                                                      maxValue: string|number|TDate|Date|DateTimeBase<TDate>|undefined|null,
                                                       valueFormat: DateTimeValueFormat|undefined|null,
                                                       stringFormat: string|undefined|null,): ValidatorFn
 {
@@ -81,7 +177,7 @@ export function datetimeMaxValidator<TDate = unknown>(dateApi: DateApi<TDate>,
 
             maxValue = value;
         }
-        
+
         const parsedValue = parseDateTime(control.value, dateApi, valueFormat, stringFormat);
 
         if(!parsedValue)
@@ -117,7 +213,7 @@ export function datetimeMaxValidator<TDate = unknown>(dateApi: DateApi<TDate>,
  * @param stringFormat - Optional string format of value
  */
 export function datetimeMinValidator<TDate = unknown>(dateApi: DateApi<TDate>,
-                                                      minValue: string|number|TDate|Date|DateTimeBase<TDate>,
+                                                      minValue: string|number|TDate|Date|DateTimeBase<TDate>|undefined|null,
                                                       valueFormat: DateTimeValueFormat|undefined|null,
                                                       stringFormat: string|undefined|null,): ValidatorFn
 {
@@ -140,7 +236,7 @@ export function datetimeMinValidator<TDate = unknown>(dateApi: DateApi<TDate>,
 
             minValue = value;
         }
-        
+
         const parsedValue = parseDateTime(control.value, dateApi, valueFormat, stringFormat);
 
         if(!parsedValue)